crypto_dashboard/requirements.txt
# FastAPI and Web
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# API Client
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
pydantic==2.5.2
python-dateutil==2.8.2

tensorflow==2.15.0
keras==2.15.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2